LPS-68118 Move MDR tests
